Refactor gestao models into package
