+++ gestao/models/__init__.py
@@ -1,0 +1,23 @@
+from .cliente import Cliente
+from .programa_fidelidade import ProgramaFidelidade
+from .aeroporto import Aeroporto
+from .conta_fidelidade import ContaFidelidade
+from .emissao_passagem import EmissaoPassagem
+from .valor_milheiro import ValorMilheiro
+from .movimentacao import Movimentacao
+from .emissao_hotel import EmissaoHotel
+from .acesso_cliente_log import AcessoClienteLog
+from .cotacao_voo import CotacaoVoo
+
+__all__ = [
+    'Cliente',
+    'ProgramaFidelidade',
+    'Aeroporto',
+    'ContaFidelidade',
+    'EmissaoPassagem',
+    'ValorMilheiro',
+    'Movimentacao',
+    'EmissaoHotel',
+    'AcessoClienteLog',
+    'CotacaoVoo',
+]
