+++ gestao/models/emissao_passagem.py
@@ -1,0 +1,42 @@
+from django.db import models
+from .cliente import Cliente
+from .programa_fidelidade import ProgramaFidelidade
+from .aeroporto import Aeroporto
+
+class EmissaoPassagem(models.Model):
+    cliente = models.ForeignKey(Cliente, on_delete=models.CASCADE)
+    programa = models.ForeignKey(
+        ProgramaFidelidade, on_delete=models.CASCADE, null=True, blank=True
+    )
+    aeroporto_partida = models.ForeignKey(
+        Aeroporto,
+        on_delete=models.CASCADE,
+        related_name="partidas",
+        null=True,
+        blank=True,
+    )
+    aeroporto_destino = models.ForeignKey(
+        Aeroporto,
+        on_delete=models.CASCADE,
+        related_name="destinos",
+        null=True,
+        blank=True,
+    )
+    data_ida = models.DateTimeField()
+    data_volta = models.DateTimeField(null=True, blank=True)
+    qtd_passageiros = models.PositiveIntegerField()
+    companhia_aerea = models.CharField(max_length=100, blank=True)
+    localizador = models.CharField(max_length=100, blank=True)
+    valor_referencia = models.DecimalField(max_digits=10, decimal_places=2)
+    valor_pago = models.DecimalField(max_digits=10, decimal_places=2)
+    pontos_utilizados = models.IntegerField(null=True, blank=True)
+    valor_referencia_pontos = models.DecimalField(
+        max_digits=10, decimal_places=2, null=True, blank=True
+    )
+    economia_obtida = models.DecimalField(
+        max_digits=10, decimal_places=2, null=True, blank=True
+    )
+    detalhes = models.TextField(blank=True)
+
+    def __str__(self):
+        return f"{self.cliente} - {self.programa} - {self.data_ida}"
